Separate structs into separate file
